requirements.txt
langchain-anthropic
panel
jupyter_bokeh 
altair 
vega_datasets 
ipykernel